[REF] shift test to extract common class
